Counter Editor: move components to own files (#4138)
